+++ WhiteLabelDemoApp/WhiteLabelDemoApp/SessionCustomViewController.h
@@ -1,0 +1,14 @@
+//
+//  SessionCustomViewController.h
+//  WhiteLabelDemoApp
+//
+//  Created by ani on 7/11/16.
+//  Copyright © 2016 LaunchKey. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface SessionCustomViewController : UIViewController
+@property (weak, nonatomic) IBOutlet UITableView *tblAuths;
+
+@end