+++ WhiteLabelDemoAppSwift/Authenticator.framework/Versions/A/Headers/AuthenticatorButton.h
@@ -1,0 +1,13 @@
+//
+//  AuthenticatorButton.h
+//  WhiteLabel
+//
+//  Created by ani on 11/10/16.
+//  Copyright © 2016 LaunchKey. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AuthenticatorButton : UIButton
+
+@end