+++ WhiteLabelDemoAppSwift/Authenticator.framework/Versions/A/Headers/AuthorizationSliderButton.h
@@ -1,0 +1,15 @@
+//
+//  AuthorizationSliderButton.h
+//  WhiteLabel
+//
+//  Created by ani on 1/23/17.
+//  Copyright © 2017 LaunchKey. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AuthorizationSliderButton : UIButton
+
+@property (nonatomic, strong) UIColor *highlihgtedStateColor UI_APPEARANCE_SELECTOR;
+
+@end