+++ WhiteLabelDemoAppSwift/SwiftWhiteLabelSDK/SessionCustomTableViewCell.swift
@@ -1,0 +1,17 @@
+//
+//  SessionCustomTableViewCell.swift
+//  WhiteLabelDemoAppSwift
+//
+//  Created by ani on 8/11/16.
+//  Copyright © 2016 LaunchKey. All rights reserved.
+//
+
+import Foundation
+
+class SessionCustomTableViewCell: UITableViewCell
+{
+    
+    @IBOutlet weak var labelAuthName: UILabel!
+    @IBOutlet weak var btnRemove: UIButton!
+    
+}